Attempt refactor
Removes Presentation and Slide concepts since those must exist on the disk anyway
