+++ src/presentation/mod.rs
@@ -1,69 +1,72 @@
-use slide::Slide;
-use crate::ui::PresentationConfig;
+use std::collections::HashSet;
 use anyhow::Context;
-use std::fs;
+use std::{env, fs};
 use std::path::{Path, PathBuf};
 use tera::Tera;
 use tracing::{debug, trace, warn};
+use crate::errors::ArgumentError;
+use crate::presentation::slide::io::find_slides;
+use crate::presentation::slide::SlideFile;
+use crate::ui::cli::{CliArgs, Commands};
+use crate::ui::conf::PresentationConfigFile;
 
 pub mod slide;
 
-#[derive(Debug)]
-pub struct Presentation {
-    /// The title of the presentation
+/// The logical representation of a presentation configuration
+#[derive(Debug, Clone)]
+pub struct PresentationConfig {
+    /// Title of the presentation
     pub title: String,
-    /// Contains the contents of the template to use for the presentation
-    pub template: String,
-    /// Contains the slides, in order, to include in the presentation
-    pub slides: Vec<Slide>,
+    /// Output directory of the presentation
+    /// Needs to exist
+    // todo: support directories that don't yet exist
+    pub output_directory: PathBuf,
+    /// Output filename of the final presentation file, with extension
+    pub output_filename: PathBuf,
+    /// Absolute path to the template file
+    pub template_file: PathBuf,
+    pub slides: Vec<SlideFile>,
 }
 
-/// Attempts to parse the PresentationConfig and read all the necessary details in
-/// producing a presentation
-impl TryFrom<PresentationConfig> for Presentation {
-    type Error = anyhow::Error;
+impl PresentationConfig {
+    /// Attempts to validate this PresentationConfig
+    /// In particular, it checks that any paths
+    /// specified are valid, and those that need to be
+    /// accessed can be accessed.
+    fn validate(&self) -> Result<(), ArgumentError> {
+        trace!("Validating PresentationConfig");
+        trace!("Checking output_file");
+        // todo:
 
-    /// Attempts to parse the PresentationConfig and read all the necessary details in
-    /// producing a presentation
-    ///
-    /// # Arguments
-    /// * `config` - The PresentationConfig to parse
-    ///
-    /// # Returns
-    /// A Presentation if the config is valid
-    ///
-    /// # Errors
-    /// - If the slides could not be read
-    /// - If the template could not be read
-    fn try_from(config: PresentationConfig) -> Result<Self, Self::Error> {
-        trace!("Attempting to parse PresentationConfig");
-        trace!("Presentation title: {}", &config.title);
-        trace!(
-            "Reading template_file at `{}`",
-            &config.output_file.display()
-        );
-        let template = fs::read_to_string(&config.template_file)?;
-        trace!("Template file read: {} bytes", template.len());
-        trace!("Reading {} slides", &config.include_files.len());
-        let slides = {
-            let mut slides = Vec::new();
-            for slide_file in config.include_files {
-                trace!("Reading slide at `{}`", slide_file.path.display());
-                let slide = Slide::try_from(slide_file)?;
-                slides.push(slide);
-            }
-            slides
-        };
-        trace!("Parsed {} slides", slides.len());
-        Ok(Self {
-            title: config.title,
-            template,
-            slides,
-        })
+        let output_file = self.output_directory.join(&self.output_filename);
+
+        // does it exist and is it a file?
+        if output_file.is_file() {
+            // if it exists, we will warn about overwriting it
+            warn!(
+                "Output file at `{}` already exists, will overwrite",
+                output_file.display()
+            );
+        }
+        trace!("Checking template_file");
+        if !self.template_file.is_absolute() {
+            return Err(ArgumentError::new(
+                "template_file".to_string(),
+                self.template_file.to_str().unwrap_or("<invalid path>"),
+                "Template file must be an absolute path".to_string(),
+            ));
+        }
+
+        if !self.template_file.is_file() {
+            return Err(ArgumentError::new(
+                "template_file".to_string(),
+                self.template_file.to_str().unwrap_or("<invalid path>"),
+                "Template file does not exist or cannot be read".to_string(),
+            ));
+        }
+        Ok(())
     }
-}
 
-impl Presentation {
     /// Renders the presentation into a string
     ///
     /// # Returns
@@ -71,18 +74,19 @@ impl Presentation {
     ///
     /// # Errors
     /// If the template engine fails to render the presentation.
-    pub fn render(&self) -> Result<String, tera::Error> {
+    fn render(&self) -> Result<String, tera::Error> {
         let mut ctx = tera::Context::new();
+        let template = fs::read_to_string(&self.template_file)?;
 
-        let slide_contents = self
-            .slides
-            .iter()
-            .map(|s| s.render())
-            .collect::<Vec<String>>();
+        // let slide_contents = self
+        //     .slides
+        //     .iter()
+        //     .map(| s| s.contents)
+        //     .collect::<Vec<String>>();
         ctx.insert("slide_title", &self.title);
-        ctx.insert("ingested_files", &slide_contents);
+        // ctx.insert("ingested_files", &slide_contents);
 
-        let result = Tera::one_off(&self.template, &ctx, false);
+        let result = Tera::one_off(&template, &ctx, false);
         trace!("Render template succeeded: {}", result.is_ok());
         result
     }
@@ -90,102 +94,103 @@ impl Presentation {
     /// Packages the presentation to a file.
     /// This will copy all local images referenced in slides into the output directory
     ///
-    /// # Arguments
-    /// * `output_dir`: The directory to place all the presentation output files into
-    ///
     /// Optionally, downloads revealJS libs and generates the zip too
-    pub fn package<P: AsRef<Path>>(&mut self, output_dir: P) -> Result<(), anyhow::Error> {
-        // todo: refactor logic here, too messy
+    pub fn package(&self) -> Result<(), anyhow::Error> {
         let output = self.render()?;
         debug!("Rendered {} bytes", output.len());
-        // todo: read the config!
-        let output_path = output_dir.as_ref().join("index.html");
+        let output_path = self.output_directory.join(&self.output_filename);
         debug!("Writing to `{}`", output_path.display());
         fs::write(&output_path, output)?;
         println!("Slides written to `{}`", output_path.display());
 
-        for slide in &mut self.slides {
-            if slide.slide_path.is_none() {
-                // todo: support images with absolute paths
-                warn!("Skipping a slide which has no path");
-                continue;
-            }
-            slide.parse();
-            let slide_path = slide.slide_path.as_ref().unwrap();
-
-            trace!("Slide is at {}", slide_path.display());
-
-            // safe to unwrap because we just parsed the slide
-            let local_images = slide.local_images.as_ref().unwrap();
-            trace!("Slide has {} local images", local_images.len());
-            if local_images.is_empty() {
-                continue;
-            }
-
-            for img in local_images {
-                let im_path = PathBuf::from(img);
-                let img_filename = im_path
-                    .file_name()
-                    .with_context(|| {
-                        format!("Could not obtain file name of {}", im_path.display())
-                    })?
-                    .to_str()
-                    .with_context(|| format!("{} is not valid UTF-8", im_path.display()))?;
-
-                debug!("Image filename is {}", img_filename);
-                if !img.starts_with("../img") {
-                    // todo: this might not work on windows
-                    warn!("This local image is not in the img/ directory (it's in `{}`) and will be skipped.", img);
-                    continue;
-                }
-                let mut img_containing_dir = im_path.strip_prefix("..")?.to_path_buf();
-                img_containing_dir.pop();
-
-                let slide_dir = &slide_path.parent().with_context(|| {
-                    format!(
-                        "Could not get parent of slide at path `{}`",
-                        slide_path.display()
-                    )
-                })?;
-                let actual_img_path = fs::canonicalize(slide_dir.join(img))?;
-                let img_dst_dir = output_dir.as_ref().join(img_containing_dir);
-                let img_dst_path = img_dst_dir.join(img_filename);
-
-                trace!("Attempting to create {}", img_dst_dir.display());
-                fs::create_dir_all(&img_dst_dir)?;
-                println!(
-                    "Copying `{}` into `{}`",
-                    actual_img_path.display(),
-                    img_dst_path.display()
-                );
-                fs::copy(actual_img_path, img_dst_path)?;
-            }
+        for slide in &self.slides {
+            todo!()
         }
         Ok(())
     }
 }
 
-#[cfg(test)]
-mod test {
-    use super::*;
+/// Attempts to convert CLI user input to PresentationConfig
+/// All paths will be converted to absolute paths with respect to the current working directory.
+/// (i.e. the directory the command was executed in)
+impl TryFrom<CliArgs> for PresentationConfig {
+    type Error = anyhow::Error;
 
-    #[test]
-    fn test_render() {
-        let slides = vec![
-            Slide::new("slide 1".to_string()),
-            Slide::new("slide 2".to_string()),
-            Slide::new("slide 3".to_string()),
-        ];
-        let ppt = Presentation {
-            title: "Test Presentation".to_string(),
-            template: "{{ slide_title }} {%for fc in ingested_files %}'{{fc}}'{%endfor%}"
-                .to_string(),
-            slides,
-        };
-        let render_result = ppt.render().unwrap();
-        assert_eq!(
-            render_result,
-            "Test Presentation 'slide 1''slide 2''slide 3'"
+    fn try_from(args: CliArgs) -> Result<Self, Self::Error> {
+        match args.command {
+            Commands::FromConfig { config_path } => {
+                let config = PresentationConfigFile::read_config_file(config_path)?;
+                Ok(Self::try_from(config)?)
+            }
+            Commands::FromCli {
+                title,
+                slide_dir,
+                template_file,
+                output_dir,
+                output_file,
+            } => {
+                trace!("Converting CLI args to PresentationConfig");
+                let cwd = fs::canonicalize(env::current_dir()?)?;
+                let slide_title = if let Some(title) = title {
+                    title
+                } else {
+                    "Untitled Presentation".to_string()
+                };
+                let slides = find_slides(&cwd.join(slide_dir))?;
+                let cfg = PresentationConfig {
+                    title: slide_title,
+                    output_directory: cwd.join(output_dir),
+                    output_filename: output_file,
+                    template_file: cwd.join(template_file),
+                    slides,
+                };
+                cfg.validate()?;
+                Ok(cfg)
+            }
+        }
+    }
+}
+
+/// Attempts to convert a PresentationConfigFile to PresentationConfig
+/// Validates and converts relative paths to absolute paths in the process
+impl TryFrom<PresentationConfigFile> for PresentationConfig {
+    type Error = anyhow::Error;
+
+    fn try_from(config: PresentationConfigFile) -> Result<Self, Self::Error> {
+        trace!("Attempting to convert PresentationConfigFile to PresentationConfig");
+        let include_files_abs_paths = config
+            .include_files
+            .iter()
+            .map(|relative_pth| {
+                config
+                    .working_directory
+                    .join(&config.slide_dir)
+                    .join(relative_pth)
+            })
+            .collect::<Vec<PathBuf>>();
+        trace!(
+            "Converted {} include_file paths to abs paths",
+            include_files_abs_paths.len()
         );
+        let slides = if include_files_abs_paths.is_empty() {
+            // let's try to search for slides
+            find_slides(&config.working_directory.join(config.slide_dir))?
+        } else {
+            let sf = include_files_abs_paths
+                .iter()
+                .map(SlideFile::read_and_parse)
+                .collect::<Result<Vec<SlideFile>, anyhow::Error>>()?;
+            sf
+        };
+
+        let cfg = PresentationConfig {
+            title: config.title,
+            output_directory: config.working_directory.join("output"), // todo: add config option to configure output dir
+            template_file: config.working_directory.join(config.template_file),
+            output_filename: config.output_file,
+            slides
+        };
+        cfg.validate()?;
+        Ok(cfg)
     }
 }
