+++ src/presentation/slide/mod.rs
@@ -1,73 +1,141 @@
 pub mod io;
 pub mod parsing;
 
+use std::cmp::Ordering;
 use std::fs;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
+use anyhow::Context;
+use crate::errors::ValidationError;
 
-use crate::presentation::slide::io::SlideFile;
-
+use crate::presentation::slide::io::is_markdown_file;
 use crate::presentation::slide::parsing::{get_local_links, grab_image_links};
 
-/// A single slide that can be used in a presentation.
-/// A slide has contents (which is in markdown)
-/// It may also contain local images
-///
-/// # Note
-/// Constructing a slide with contents does not automatically fill in local_images, and
-/// instead the `parse` method must be used to do so
-// todo: consider automatically parsing local images when reading a slide
-// todo: investigate whether adding serialize and deserialize is better than having a render method.
-#[derive(Debug)]
-pub struct Slide {
+
+/// A SlideFile is a slide that exists as a file on the disk somewhere
+#[derive(PartialEq, Debug, Clone)]
+pub struct SlideFile {
+    filename: String,
+    /// Absolute path to where this slideFile is located on the disk
+    pub path: PathBuf,
+    /// Full contents of the SlideFile
     pub contents: String,
-    /// Path to the slide file.
-    /// May be `None` if the slide was not read from a file.
-    pub slide_path: Option<PathBuf>,
-    /// `None` means that `contents` has not been parsed yet
-    /// If this is `Some`, the local_images may still be empty if there are no local images
-    pub local_images: Option<Vec<String>>,
+
+    pub local_images: Vec<PathBuf>
 }
 
-impl TryFrom<SlideFile> for Slide {
-    type Error = anyhow::Error;
-
-    /// This consumes the SlideFile
-    fn try_from(slide_file: SlideFile) -> Result<Self, Self::Error> {
-        let contents = fs::read_to_string(&slide_file.path)?;
-        Ok(Self {
-            contents,
-            slide_path: Some(slide_file.path),
-            local_images: None,
-        })
+impl PartialOrd for SlideFile {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        // todo: natural sorting
+        self.filename.partial_cmp(&other.filename)
     }
 }
 
-impl Slide {
-    /// Creates a new `Slide` from contents
-    /// This does not parse the contents automatically.
-    pub fn new(contents: String) -> Self {
-        Self {
-            contents,
-            slide_path: None,
-            local_images: None,
-        }
+impl Ord for SlideFile {
+    fn cmp(&self, other: &Self) -> Ordering {
+        self.filename.cmp(&other.filename)
     }
+}
 
-    /// Parses the contents of the slide, looking for
-    /// local images and fills in the `local_images` field
-    pub fn parse(&mut self) {
-        let im_links = grab_image_links(&self.contents);
-        let local_links = get_local_links(im_links);
-        self.local_images = Some(local_links);
-    }
+impl Eq for SlideFile {}
 
-    /// Renders the slide
+impl SlideFile {
+    /// Reads a SlideFile from the disk.
+    /// This will also transform any local links to be relative to <OUTPUT_DIR>/img/
     ///
-    /// # Note
-    /// This actually just clones the contents, which is technically inefficient.
-    /// We will see if this becomes a problem.
-    pub fn render(&self) -> String {
-        self.contents.clone()
+    /// For example, say you have a slide with the following content
+    /// ```markdown
+    /// ![](../whatever/image.png)
+    /// ```
+    /// When read in, the `content` of this SlideFile will contain instead
+    /// ```markdown
+    /// ![](./img/whatever/image.png)
+    /// ```
+    ///
+    ///
+    /// # Arguments
+    /// * `path` - Absolute path to the SlideFile on the disk
+    ///
+    /// # Errors
+    /// * `ValidationError` - If the SlideFile is not a valid SlideFile
+    /// * `std::io::Error` - If there was an error reading the SlideFile
+    ///
+    /// # Notes
+    /// This is a blocking operation since it will read the file from the disk
+    /// and attempt to parse it.
+    pub fn read_and_parse<P: AsRef<Path>>(path: P) -> Result<Self, anyhow::Error> {
+        let filename = path
+            .as_ref()
+            .file_name()
+            .with_context(|| format!("`{}` does not contain a valid filename", path.as_ref().display()))?
+            .to_str()
+            .with_context(|| format!("Filename at `{}` is not UTF-8!", path.as_ref().display()))?
+            .to_string();
+        SlideFile::validate_path(&path)?;
+        let contents = fs::read_to_string(&path)?;
+
+        let local_images = get_local_links(&grab_image_links(&contents));
+
+        let sf = Self { filename, path: path.as_ref().to_path_buf(),
+            contents, local_images };
+        Ok(sf)
+    }
+
+    /// Creates a list of SlideFiles from paths
+    /// # Arguments
+    /// * `paths` - A list of paths to slide files.
+    ///
+    /// # Returns
+    /// A list of SlideFiles.
+    ///
+    /// # Errors
+    /// - If a slide file has an invalid file name
+    /// - If a slide file has a filename that is not UTF-8 compatible
+    fn from_paths(paths: Vec<PathBuf>) -> Result<Vec<Self>, anyhow::Error> {
+        paths
+            .into_iter()
+            .map(SlideFile::read_and_parse)
+            .collect::<Result<Vec<SlideFile>, anyhow::Error>>()
+    }
+
+    /// Attempts to validate the path to a SlideFile
+    /// This checks if the file
+    /// - actually exists
+    /// - can be read
+    /// - is a .md file
+    ///
+    /// # Returns
+    /// None
+    ///
+    /// # Errors
+    /// - If the slide file does not exist
+    /// - If the slide file is not a file
+    /// - If the slide file is not a markdown file
+    fn validate_path<P: AsRef<Path>>(slide_file_path: P) -> Result<(), ValidationError> {
+        if !slide_file_path.as_ref().is_absolute() {
+            return Err(ValidationError::new(
+                &slide_file_path.as_ref().display().to_string(),
+                "Path is not absolute".to_string(),
+            ));
+        }
+        if !slide_file_path.as_ref().exists() {
+            return Err(ValidationError::new(
+                &slide_file_path.as_ref().display().to_string(),
+                "File does not exist".to_string(),
+            ));
+        }
+        if !slide_file_path.as_ref().is_file() {
+            return Err(ValidationError::new(
+                &slide_file_path.as_ref().display().to_string(),
+                "Path is not a file".to_string(),
+            ));
+        }
+        if !is_markdown_file(&slide_file_path.as_ref()) {
+            return Err(ValidationError::new(
+                &slide_file_path.as_ref().display().to_string(),
+                "File is not a markdown file".to_string(),
+            ));
+        }
+        Ok(())
     }
 }
 
@@ -79,33 +147,22 @@ mod test {
     use tempfile::tempdir;
 
     #[test]
-    fn test_slide_from_file() {
-        let slide_contents = r#"
-# Hello World
-
-This is a markdown slide, used for testing.
-        "#;
+    fn test_parse_slide() {
+        let slide_contents = r#"![oh no an image](./local/image.png)"#;
         let tmp_dir = tempdir().unwrap();
-        let slide_file = tmp_dir.path().join("slide.md");
+        let abs_path_to_tmp_dir = fs::canonicalize(tmp_dir.path()).unwrap();
+        let slide_file = abs_path_to_tmp_dir.join("slide.md");
         let mut h_slide_file = File::create(&slide_file).unwrap();
         h_slide_file.write_all(slide_contents.as_bytes()).unwrap();
-        let s_slidefile = SlideFile::try_from(slide_file).unwrap();
-        let slide = Slide::try_from(s_slidefile).unwrap();
-        assert_eq!(slide.contents, slide_contents);
-    }
 
-    #[test]
-    fn test_slide_creation_does_not_auto_parse_images() {
-        let slide_contents = "Cool little slide eh ![oh no an image](/haha/local/image.png)";
-        let slide = Slide::new(slide_contents.to_string());
-        assert_eq!(slide.local_images, None);
-    }
+        let local_img = abs_path_to_tmp_dir.join("local/image.png");
+        fs::create_dir_all(local_img.parent().unwrap()).unwrap();
+        let h_local_img = File::create(&local_img).unwrap();
 
-    #[test]
-    fn test_slide_parsing() {
-        let slide_contents = "Cool little slide eh ![oh no an image](/haha/local/image.png)";
-        let mut slide = Slide::new(slide_contents.to_string());
-        slide.parse();
-        assert_eq!(slide.local_images.unwrap()[0], "/haha/local/image.png");
+        let slide_file = SlideFile::read_and_parse(slide_file).unwrap();
+        assert_eq!(slide_file.contents, slide_contents);
+        assert_eq!(slide_file.local_images.len(), 1);
+        assert_eq!(slide_file.local_images[0],
+                   (PathBuf::from("/haha/local/image.png"), PathBuf::from("")));
     }
 }
